airflow_project/requirements.txt
sqlalchemy
pandas
numpy
azure-storage-blob
requests
beautifulsoup4
fake-useragent
python-dotenv
